+++ Sources/SwiftFortuneWheel/Utils/Drawing/SliceDrawing.swift
@@ -189,16 +189,13 @@ extension SliceDrawing {
         switch preferences.orientation {
         case .horizontal:
             if preferences.isCurved {
-                self.drawCurved(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
+                return self.drawCurved(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
             } else {
-                self.drawHorizontal(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
+                return self.drawHorizontal(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
             }
         case .vertical:
-            // Not finished yet
-//            self.drawVertical(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
-            self.drawHorizontal(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
+            return self.drawVertical(text: text, in: context, preferences: preferences, rotation: rotation, index: index, topOffset: topOffset)
         }
-        return preferences.preferedFontSize + preferences.verticalOffset
     }
 
     /// Draws curved text
@@ -209,20 +206,13 @@ extension SliceDrawing {
     ///   - rotation: rotation degree
     ///   - index: index
     ///   - topOffset: top offset
-    private func drawCurved(text: String, in context:CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) {
+    private func drawCurved(text: String, in context:CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) -> CGFloat {
 
-        var textColor: UIColor = .black
+        let textColor = preferences.color(for: index)
 
-        switch preferences.textColorType {
-        case .evenOddColors(let evenColor, let oddColor):
-            textColor = index % 2 == 0 ? evenColor : oddColor
-        case .customPatternColors(let colors, let defaultColor):
-            textColor = colors?[index, default: defaultColor] ?? defaultColor
-        }
-
-        let bottomYPosition = -(radius - preferences.verticalOffset - topOffset - preferences.preferedFontSize)
+        let bottomYPosition = -(radius - preferences.verticalOffset - topOffset - preferences.font.pointSize)
         let width = self.width(forYPosition: bottomYPosition)
-        let textRect = CGRect(x: 0, y: 0, width: width, height: preferences.preferedFontSize)
+        let textRect = CGRect(x: 0, y: 0, width: width, height: preferences.font.pointSize)
         let yPosition = -(radius - preferences.verticalOffset - topOffset - topMargin) + textRect.height / 2
 
         context.saveGState()
@@ -233,6 +223,8 @@ extension SliceDrawing {
 
         context.restoreGState()
 
+        return preferences.font.pointSize + preferences.verticalOffset
+
     }
 
     /// Draws text
@@ -243,42 +235,31 @@ extension SliceDrawing {
     ///   - rotation: rotation degree
     ///   - index: index
     ///   - topOffset: top offset
-    private func drawHorizontal(text: String, in context:CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) {
+    private func drawHorizontal(text: String, in context:CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) -> CGFloat {
 
-        var textColor: UIColor = .black
+        let textFontAttributes = preferences.textFontAttributes(for: index)
 
-        switch preferences.textColorType {
-        case .evenOddColors(let evenColor, let oddColor):
-            textColor = index % 2 == 0 ? evenColor : oddColor
-        case .customPatternColors(let colors, let defaultColor):
-            textColor = colors?[index, default: defaultColor] ?? defaultColor
-        }
-
-        let textFontAttributes: [NSAttributedString.Key:Any] = {
-            let textStyle = NSMutableParagraphStyle()
-            textStyle.alignment = .center
-            let deafultAttributes:[NSAttributedString.Key: Any] =
-                [.font: preferences.font,
-                 .foregroundColor: textColor,
-                 .paragraphStyle: textStyle ]
-            return deafultAttributes
-        }()
-
-        let bottomYPosition = -(radius - preferences.verticalOffset - topOffset - preferences.preferedFontSize)
+        let bottomYPosition = -(radius - preferences.verticalOffset - topOffset - preferences.font.pointSize)
         let width = self.width(forYPosition: bottomYPosition)
-        let textRect = CGRect(x: 0, y: 0, width: width, height: preferences.preferedFontSize)
+        let textRect = CGRect(x: 0, y: 0, width: width, height: preferences.font.pointSize)
         let yPosition = -(radius - preferences.verticalOffset - topOffset - topMargin)
+        let xPos = -(textRect.width / 2) - preferences.horizontalOffset
 
         context.saveGState()
         context.rotate(by: rotation * CGFloat.pi/180)
 
-        if preferences.flipUpsideDown {
+        context.translateBy(x: xPos, y: yPosition)
+
+        if !preferences.flipUpsideDown {
             context.rotate(by: flipRotation)
+            context.translateBy(x: -textRect.width, y: -textRect.height)
         }
 
-        text.draw(in: CGRect(x: -(textRect.width / 2) - preferences.horizontalOffset, y: yPosition, width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
+        text.draw(in: CGRect(x: 0, y: 0, width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
 
         context.restoreGState()
+
+        return preferences.font.pointSize + preferences.verticalOffset
 
     }
 
@@ -290,61 +271,46 @@ extension SliceDrawing {
     ///   - rotation: rotation degree
     ///   - index: index
     ///   - topOffset: top offset
-    private func drawVertical(text: String, in context: CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) {
+    private func drawVertical(text: String, in context: CGContext, preferences: TextPreferences, rotation: CGFloat, index: Int, topOffset: CGFloat) -> CGFloat {
 
-        var textColor: UIColor = .black
-
-        switch preferences.textColorType {
-        case .evenOddColors(let evenColor, let oddColor):
-            textColor = index % 2 == 0 ? evenColor : oddColor
-        case .customPatternColors(let colors, let defaultColor):
-            textColor = colors?[index, default: defaultColor] ?? defaultColor
-        }
-
-        let textFontAttributes: [NSAttributedString.Key:Any] = {
-            let textStyle = NSMutableParagraphStyle()
-            textStyle.alignment = .center
-            let deafultAttributes:[NSAttributedString.Key: Any] =
-                [.font: preferences.font,
-                 .foregroundColor: textColor,
-                 .paragraphStyle: textStyle ]
-            return deafultAttributes
-        }()
+        let textFontAttributes = preferences.textFontAttributes(for: index)
 
         let textWidth: CGFloat = text.width(by: preferences.font)
 
-        let maxAvailableBottomRadiusOffsetCircularSegmentHeight = self.leftMargin + self.rightMargin + preferences.preferedFontSize
+        let maxAvailableBottomRadiusOffsetCircularSegmentHeight = self.leftMargin + self.rightMargin + preferences.font.pointSize
         let bottomRadiusOffset = max(self.bottomMargin, radius(circularSegmentHeight: maxAvailableBottomRadiusOffsetCircularSegmentHeight, from: sliceDegree))
 
         let availableHeightInSlice: CGFloat = radius - preferences.verticalOffset - topOffset - bottomRadiusOffset
         let textRectWidth = min(textWidth, availableHeightInSlice)
         let croppedText = text.crop(by: textRectWidth, font: preferences.font)
 
-//        let bottomYPosition = -(radius - preferences.verticalOffset - topOffset - preferences.preferedFontSize)
-        let textRect = CGRect(x: 0, y: 0, width: textRectWidth, height: preferences.preferedFontSize)
+        let textRect = CGRect(x: 0, y: 0, width: textRectWidth, height: preferences.font.pointSize)
         let yPosition = -(radius - preferences.verticalOffset - topOffset - topMargin)
-        let xPos = -(textRect.width / 2) - preferences.horizontalOffset
+        let xPos = -(textRect.height / 2) - preferences.horizontalOffset
 
         context.saveGState()
         context.rotate(by: rotation * CGFloat.pi/180)
-        context.translateBy(x: xPos + textRect.height / 2, y: yPosition + textRect.width / 2)
+
+        context.translateBy(x: -xPos, y: yPosition)
         context.rotate(by: 90 * CGFloat.pi/180)
 
 
-//        context.rotate(by: 90 * CGFloat.pi/180)
-
         if preferences.flipUpsideDown {
             context.rotate(by: flipRotation)
+            context.translateBy(x: -textRectWidth, y: -textRect.height)
         }
 
-//        croppedText.draw(in: CGRect(x: 0, y: 0, width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
-//        croppedText.draw(in: CGRect(x: -(textRect.width / 2) - preferences.horizontalOffset - rotationOffset / 2, y: yPosition - rotationOffset / 2 , width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
-//        context.addRect(textRect)
-//        UIColor.red.setStroke()
-//        context.drawPath(using: .fillStroke)
-        croppedText.draw(in: CGRect(x: -textRect.width / 2, y: -textRect.height / 2, width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
+
+        // For Debugging purposes
+        //        context.addRect(textRect)
+        //        UIColor.red.setStroke()
+        //        context.drawPath(using: .fillStroke)
+
+        croppedText.draw(in: CGRect(x: 0, y: 0, width: textRect.width, height: textRect.height), withAttributes: textFontAttributes)
 
         context.restoreGState()
+
+        return textRectWidth + preferences.verticalOffset
 
     }
 }
@@ -377,17 +343,10 @@ extension SliceDrawing {
     ///   - topOffset: top offset
     func drawLine(in context: CGContext, preferences: LinePreferences, start: CGFloat, and end: CGFloat, rotation: CGFloat, index: Int, topOffset: CGFloat) {
 
+        let strokeColor = preferences.strokeColor(for: index)
+
         context.saveGState()
         context.rotate(by: (rotation - contextPositionCorrectionOffsetDegree) * CGFloat.pi/180)
-
-        var strokeColor = UIColor.clear
-
-        switch preferences.colorType {
-        case .evenOddColors(let evenColor, let oddColor):
-            strokeColor = index % 2 == 0 ? evenColor : oddColor
-        case .customPatternColors(let colors, let defaultColor):
-            strokeColor = colors?[index, default: defaultColor] ?? defaultColor
-        }
 
         let strokeWidth = preferences.height